+++ src/TimeFormatLength.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Bakame\Tokei;
+
+enum TimeFormatLength
+{
+    case Short;
+    case Medium;
+    case Long;
+    case Full;
+}